Move mod core to file
